// client/PC-World/src/app/shared/interfaces/notebook.interface.ts
export interface INotebook {
    _id: string;
    images: string[];
    brand: string;
    model: string;
    processor: string;
    processorCores: number | string;
    memoryType: string;
    memoryCapacity: number | string;
    memorySpeed: number | string;
    storage: string;
    storageCapacity: number | string;
    graphics: string;
    display: string;
    displaySize: number | string;
    displayResolution: string;
    displayRefreshRate: number | string;
    ports: string;
    weight: number | string;
    color: string;
    dimensions: string;
    battery: string;
    OS: string;
    category: string;
    warranty: number | string;
    price: number | string;
    quantity: number | string;
    processorBrand: string;
    processorModel: string;
    promoPrice: number | string;
}